Fix #139 computation of missing lat/lon for GAC. Move _compute_missing_lonlat to baseclass and use pyorbital avhrr*_from_times if available.
